+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * `delete_link` now adds a `data-confirm` attribute by default (#25)
 * `{details,edit,delete}_link` integrate with authorisation, if possible (#26)
 
+### Added
+* Add `inline_controls_for` button toolbar helper (#27)
+
 ## 1.12.2 / 2018-06-22
 ### Fixed
 * Address issue with datepicker SCSS (#22)